Aplicacion de vista rating y desarrollo de urls
