+++ app/templates/app/event_rating.html
@@ -3,7 +3,6 @@
 {% block content %}
 <div class="container">
     <div class="row">
-        
         {% if messages %}
             {% for message in messages%}
                 <div class="alert alert-success" role="alert">
@@ -53,39 +52,40 @@
             </div>
         </div>
     </div>
-    <hr>
+    <div class="container">
+        <h2 class="mb-4">Calificaciones y Reseñas</h2>
+            {% for rating in ratings %}
+                <div class="card mb-4 shadow-sm">
+                    <div class="card-body">
+                        <h5 class="card-title">{{ rating.usuario.username }}</h5>
+                        <h6 class="card-subtitle mb-2 text-muted">{{ rating.fecha_creacion|date:"d/m/Y H:i" }}</h6>
+                        <p class="card-text"><strong>{{ rating.titulo }}</strong></p>
+                        <p class="card-text">{{ rating.calificacion }} ⭐</p>
+                        <p class="card-text">{{ rating.texto }}</p>
 
-    {% if user.is_authenticated %}
-        <h2>{% if editando %}Editar tu reseña{% else %}Calificar este evento{% endif %}</h2>
-        <form method="post">
-            {% csrf_token %}
-            {{ form.as_p }}
-            <button type="submit" class="btn btn-primary">
-                {% if editando %}Actualizar reseña {%else%} Enviar reseña {% endif %}
-            </button>
-        </form>
-    {%else%%}
-        <p><a href="{% url 'login' %}">Iniciar sesion</a> para modificar este evento</p>
-    {%endif%%}
+                        {% if user.is_authenticated and user == rating.usuario or user.is_staff %}
+                            <a href="{% url 'editar_rating' rating.id %}" class="btn btn-sm btn-warning">Editar</a>
+                            <a href="{% url 'eliminar_rating' rating.id %}" class="btn btn-sm btn-danger">Eliminar</a>
+                        {% endif %}
 
-    <h2>Calificaciones y Reseñas</h2>
-        {% for rating in ratings %}
-            <div class="card mb-3">
-                <div class="card-body">
-                    <h5 class="card-title">{{ rating.usuario.username }}</h5>
-                    <h6 class="card-subtitle mb-2 text-muted">{{ rating.fecha_creacion|date:"d/m/Y H:i" }}</h6>
-                    <p class="card-text"><strong>{{ rating.titulo }}</strong></p>
-                    <p class="card-text">{{ rating.calificacion }} ⭐</p>
-                    <p class="card-text">{{ rating.texto }}</p>
-
-                    {% if user.is_authenticated and (user == rating.usuario or user.is_staff) %}
-                        <a href="{% url 'editar_rating' rating.id %}" class="btn btn-sm btn-warning">Editar</a>
-                        <a href="{% url 'eliminar_rating' rating.id %}" class="btn btn-sm btn-danger">Eliminar</a>
-                    {% endif %}
+                    </div>
                 </div>
-            </div>
-        {% empty %}
-            <p>Aún no hay reseñas para este evento. ¡Sé el primero en opinar!</p>
-        {% endfor %}
+            {% empty %}
+                <p>Aún no hay reseñas para este evento. ¡Sé el primero en opinar!</p>
+            {% endfor %}
+        
+            {% if user.is_authenticated %}
+            <h2>{% if editando %}Editar tu reseña{% else %}Calificar este evento{% endif %}</h2>
+            <form method="post">
+                {% csrf_token %}
+                {{ form.as_p }}
+                <button type="submit" class="btn btn-primary">
+                    {% if editando %}Actualizar reseña {%else%} Enviar reseña {% endif %}
+                </button>
+            </form>
+            {% else %}
+                <p><a href="{% url 'login' %}">Iniciar sesion</a> para modificar este evento</p>
+            {% endif %}
+    </div>
 </div>
 {% endblock %}
