+++ app/urls.py
@@ -12,5 +12,6 @@ urlpatterns = [
     path("events/create/", views.event_form, name="event_form"),
     path("events/<int:id>/edit/", views.event_form, name="event_edit"),
     path("events/<int:id>/", views.event_detail, name="event_detail"),
+    path("events/<int:id>/rating", views.detalle_evento, name ="event_rating"),
     path("events/<int:id>/delete/", views.event_delete, name="event_delete"),
 ]
